backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and security
firebase-admin==6.2.0
pyjwt==2.8.0
cryptography==41.0.7

# ML and AI
tensorflow==2.15.0
tensorflow-lite==2.15.0
mediapipe==0.10.7
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Image and video processing
pillow==10.1.0
imageio==2.33.0
moviepy==1.0.3

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
